requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
djangorestframework==3.6.3
